优化 android 像素单位处理

+++ android/src/main/java/cn/qiuxiang/react/amap3d/AMapUtilsModule.kt
@@ -20,5 +20,5 @@ class AMapUtilsModule(reactContext: ReactApplicationContext) : ReactContextBaseJ
     }
 }
 
-val Int.toPx: Int
+val Float.toPx: Int
     get() = (this * Resources.getSystem().displayMetrics.density).toInt()
